==> 04 Train Set Creation (Mordred).ipynb <==
### Imports

import pandas as pd
import numpy as np
import pickle

### Loading Data

mordred_all_active = pd.read_csv("./data/0_all_active_all_inactive/Mordred/all_active_mordred.csv")
mordred_all_inactive = pd.read_csv("./data/0_all_active_all_inactive/Mordred/all_inactive_mordred.csv")
mordred_366_random_inactive = pd.read_csv("./data/1_all_active_random_inactive/Mordred/random_inactive_mordred.csv")
with open('./data/2_all_active_autoencoder_inactive/Mordred/mordred_most_dissimilar_compounds_366.pickle', 'rb') as file5:
    mordred_autoencoder_selected_366_inactive = pickle.load(file5)

def get_ext_int_train(initial_active, initial_inactive):
    """
    Accepts: two dataframes with descriptor values for active and inactive compounds
             (already with the target column)
    Returns: ext_x_test, ext_y_test, int_x_test, int_y_test, x_train, y_train
    ! ! ! Note that its sample splits are hardcorded ! ! ! 
    """
    external_active = initial_active.sample(n=36, random_state=42)
    external_inactive = initial_inactive.sample(n=36, random_state=42)
    
    external = pd.concat([external_active, external_inactive])
    
    external_y = external["target"].values # returned
    external_x = external.drop(["target"], axis=1).values # returned
    
    remaining_active = initial_active.drop(index=external_active.index)
    remaining_inactive = initial_inactive.drop(index=external_inactive.index)
    
    internal_active = remaining_active.sample(n=33, random_state=42)
    internal_inactive = remaining_inactive.sample(n=33, random_state=42)
    
    internal = pd.concat([internal_active, internal_inactive])
    
    internal_y = internal["target"].values # returned
    internal_x = internal.drop(["target"], axis=1).values # returned
    
    lastly_remaining_active = remaining_active.drop(index=internal_active.index)
    lastly_remaining_inactive = remaining_inactive.drop(index=internal_inactive.index)
    
    train = pd.concat([lastly_remaining_active, lastly_remaining_inactive])
    
    train_y = train["target"].values # returned
    train_x = train.drop(['target'], axis=1).values #returned 
        
    return external_x, external_y, internal_x, internal_y, train_x, train_y

mordred_all_active["target"] = 1

mordred_all_inactive["target"] = 0

mordred_366_random_inactive["target"] = 0

mordred_366_random_inactive.columns

mordred_autoencoder_selected_366_inactive_df = pd.DataFrame(data=mordred_autoencoder_selected_366_inactive)

mordred_autoencoder_selected_366_inactive_df["target"] = 0

mordred_autoencoder_selected_366_inactive_df.columns = mordred_366_random_inactive.columns

mordred_autoencoder_selected_366_inactive_df.shape

external_x, external_y, internal_x, internal_y, train_x, train_y = get_ext_int_train(mordred_all_active, mordred_autoencoder_selected_366_inactive_df)




external_x

external_x.shape

with open("./data/2_all_active_autoencoder_inactive/Mordred/for_ml/external_test_set/external_x_test.pickle", "wb") as file1:
    pickle.dump(external_x, file1)
    
with open("./data/2_all_active_autoencoder_inactive/Mordred/for_ml/external_test_set/external_y_test.pickle", "wb") as file2:
    pickle.dump(external_y, file2)
    
with open("./data/2_all_active_autoencoder_inactive/Mordred/for_ml/internal_test_set/internal_x_test.pickle", "wb") as file3:
    pickle.dump(internal_x, file3)
    
with open("./data/2_all_active_autoencoder_inactive/Mordred/for_ml/internal_test_set/internal_y_test.pickle", "wb") as file4:
    pickle.dump(internal_y, file4)
    
with open("./data/2_all_active_autoencoder_inactive/Mordred/for_ml/train_set/x_train.pickle", "wb") as file5:
    pickle.dump(train_x, file5)

with open("./data/2_all_active_autoencoder_inactive/Mordred/for_ml/train_set/y_train.pickle", "wb") as file6:
    pickle.dump(train_y, file6)
